refactor(frontend): split 2-(pre)fetch related code to a separate file
